Updated module augmentation for html-support package.
