+++ packages/ckeditor5-html-support/src/generalhtmlsupportconfig.ts
@@ -71,17 +71,3 @@ export interface GeneralHtmlSupportConfig {
 	disallow?: Array<MatcherPattern>;
 
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface EditorConfig {
-
-		/**
-		 * The configuration of the General HTML Support feature.
-		 * Introduced by the {@link module:html-support/generalhtmlsupport~GeneralHtmlSupport} feature.
-		 *
-		 * Read more in {@link module:html-support/generalhtmlsupportconfig~GeneralHtmlSupportConfig}.
-		 */
-		htmlSupport?: GeneralHtmlSupportConfig;
-	}
-
-}
